examples/v0.4/MicroBlock/v1.0/internal_ftp/requirements.txt
pyftpdlib==2.0.1
pyserial==3.5
